+++ src/tx-data/sell-all.js
@@ -17,6 +17,13 @@ class MinterTxDataSellAll {
                 allowZero: true,
                 length: 10,
                 default: new Buffer([]),
+            },
+            {
+                name: 'minimumValueToBuy',
+                length: 32,
+                allowZero: true,
+                allowLess: true,
+                default: new Buffer([]),
             }];
 
         /**