+++ src/tx-data/sell.js
@@ -24,6 +24,13 @@ class MinterTxDataSell {
                 allowZero: true,
                 length: 10,
                 default: new Buffer([]),
+            },
+            {
+                name: 'minimumValueToBuy',
+                length: 32,
+                allowZero: true,
+                allowLess: true,
+                default: new Buffer([]),
             }];
 
         /**