+++ test/tx-data/sell.test.js
@@ -6,18 +6,20 @@ describe('MinterTxDataSell', () => {
     test('rlp encoded fields', () => {
         const serializedTxData = (new MinterTxDataSell({
             coinToSell: formatCoin('MNT'),
-            valueToSell: 10,
+            valueToSell: 1000,
             coinToBuy: formatCoin('BELTCOIN'),
+            minimumValueToBuy: 5,
         })).serialize();
 
         expect(serializedTxData)
-            .toEqual(Buffer.from([215, 138, 77, 78, 84, 0, 0, 0, 0, 0, 0, 0, 10, 138, 66, 69, 76, 84, 67, 79, 73, 78, 0, 0]));
+            .toEqual(Buffer.from([218, 138, 77, 78, 84, 0, 0, 0, 0, 0, 0, 0, 130, 3, 232, 138, 66, 69, 76, 84, 67, 79, 73, 78, 0, 0, 5]));
 
         expect(decodeToArray(serializedTxData))
             .toEqual([
                 [77, 78, 84, 0, 0, 0, 0, 0, 0, 0],
-                [10],
+                [3, 232],
                 [66, 69, 76, 84, 67, 79, 73, 78, 0, 0],
+                [5],
             ]);
     });
 });